backend/requirements.txt
django
psycopg2
python-dotenv
djangorestframework
djangorestframework-simplejwt
